+++ process-manager/examples/send_signal.rs
@@ -1,7 +1,7 @@
-use smearor_wrot_process::ProcessConfig;
-use smearor_wrot_process::ProcessManager;
-use smearor_wrot_process::Signal;
-use smearor_wrot_process::StdioConfig;
+use process_manager::ProcessConfig;
+use process_manager::ProcessManager;
+use process_manager::Signal;
+use process_manager::StdioConfig;
 
 fn main() {
     let manager = ProcessManager::new();